ahmad - changeset - table moved inside body dir current view/body/table/index.tsx - v60
