+++ lib/ui/screen-components/protected/coupons/view/body/table/index.tsx
