+++ lib/ui/screen-components/protected/cuisines/view/body/table/index.tsx
